+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+#### [0.4.1] - 2016-18-01
+* fixed crash for empty `chartValues`
+* fixed `removePieceAtIndex:` if animations disabled
+* implemented some simple Tests
+* travis updated
+
 #### [0.4.0] - 2016-17-01
 * `chartValues` moved to private, added `setChartValues:` instead
 * `setValue:forIndex:` renamed to `setValue:pieceAtIndex:`